microservices/classroom_notification_service/requirements.txt
google-cloud-pubsub==2.13.11
google-cloud-bigquery==3.4.1
google-cloud-storage==1.42.2
google-cloud-secret-manager==2.12.6
google-auth==2.15.0
google-api-python-client==2.69.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
protobuf==3.20.3
oauth2client==4.1.3
ttl-cache==1.6